Rename transaction to handler
