--- packages/transactions/source/handlers/index.ts
@@ -1,2 +1,0 @@
-export type { TransactionHandlerConstructor } from "./transaction.js";
-export { TransactionHandler } from "./transaction.js";
