+++ packages/transactions/source/index.ts
@@ -1,2 +1,2 @@
-export * as Handlers from "./handlers/index.js";
+export * from "./handler.js";
 export * from "./service-provider.js";
